HT_22/django_scrapper/apps/cart/serializers.py
from rest_framework import serializers



class CartSerializer(serializers.Serializer):
    product_id = serializers.CharField()
    price = serializers.FloatField()
    quantity = serializers.IntegerField()