HT_22/requirements.txt
Django==5.0.1
fake-useragent==1.4.0
requests==2.31.0
djangorestframework==3.14.0
